refactor(v0.74.0): llm/router.py 1046L → 14-file 2-level 패키지 분할 (Tier 3 #6) (#903)
Tier 3 God Object 분할 시리즈 여섯 번째. 모든 Anthropic/OpenAI/GLM/Codex
호출의 중앙 dispatcher인 1046 LOC monolith를 두 단계 패키지로 분할.

새 패키지 구조 — core/llm/router/:
  Top level (6 files, 450L):
    __init__.py    169L  ~50개 심볼 re-export (외부 41곳 import 변경 0)
    _hooks.py       37L  _hooks_ctx, set_router_hooks, _fire_hook
    tracing.py      45L  is_langsmith_enabled, maybe_traceable
    _usage.py       74L  _record_response_usage, _record_openai_usage
    models.py       33L  ToolCallRecord, ToolUseResult dataclasses
    _di.py          92L  5 ContextVar + 6 accessor

  Sub-package calls/ (8 files, 913L):
    __init__.py     28L  re-exports
    _route.py       40L  _route_provider
    _failover.py   143L  call_with_failover
    text.py        129L  call_llm
    parsed.py      140L  call_llm_parsed
    json.py         68L  call_llm_json
    tools.py       228L  call_llm_with_tools (최대 leaf)
    streaming.py   137L  call_llm_streaming

최대 단일 파일: 1046 → 228 LOC (-78%, 다른 split과 동등한 임팩트).

테스트 호환성 — 17 @patch + 1 monkeypatch 사이트 leaf path 마이그레이션:
- tests/test_failover.py: 8곳 (call_llm × 6 → calls.json, get_anthropic_client × 2 → calls.text)
- tests/test_tool_use.py: 4곳 (calls.tools.get_anthropic_client)
- tests/test_llm_client.py: 11곳 (calls.{parsed,text,tracing}.X)
- tests/test_routing_policy.py: monkeypatch _resolve_provider → calls._route, inspect.getsource → pkgutil.iter_modules walk

테스트 5파일 (claude_adapter, goal_decomposer, native_tools,
anthropic_sampling_params, agentic_loop) 12+ 패치는 __init__.py
re-export 통해 변경 없이 작동 — 호출자가 lazy 패키지 임포트 사용 시.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (301 source files, +13)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- E2E Cowboy Bebop A (68.4) 변동 없음
- 외부 import 41곳 변경 0
- 함수 본문 byte-identical

남은 Tier 3 후보 3개 — commands.py(2441), cli/__init__.py(1889),
agent/loop.py(1754).
