--- core/llm/router.py
@@ -1,1046 +1,0 @@
-"""LLM Router — provider-aware dispatching for all LLM calls.
-
-Replaces the monolithic client.py. Routes to the correct provider SDK
-based on model name. Owns ContextVar DI definitions and LangSmith tracing.
-
-Data types (ToolCallRecord, ToolUseResult) and re-exports from
-token_tracker are kept here for backward compatibility.
-"""
-
-from __future__ import annotations
-
-import json
-import logging
-import os as _os
-import random
-import re
-import time
-from collections.abc import Iterator
-from contextvars import ContextVar
-from dataclasses import dataclass
-from typing import Any, TypeVar
-
-from pydantic import BaseModel
-
-from core.config import (
-    _resolve_provider,
-    is_model_allowed,
-    settings,
-)
-from core.hooks.utils import fire_hook
-from core.llm.adapters import CROSS_PROVIDER_FALLBACK as CROSS_PROVIDER_FALLBACK
-from core.llm.adapters import AgenticLLMPort as AgenticLLMPort
-from core.llm.adapters import ClaudeAdapter as ClaudeAdapter
-from core.llm.adapters import LLMClientPort as LLMClientPort
-from core.llm.adapters import LLMJsonCallable as LLMJsonCallable
-from core.llm.adapters import LLMParsedCallable as LLMParsedCallable
-from core.llm.adapters import LLMTextCallable as LLMTextCallable
-from core.llm.adapters import LLMToolCallable as LLMToolCallable
-from core.llm.adapters import resolve_agentic_adapter as resolve_agentic_adapter
-from core.llm.errors import (
-    LLMAPIStatusError as LLMAPIStatusError,
-)
-from core.llm.errors import (
-    LLMAuthenticationError as LLMAuthenticationError,
-)
-from core.llm.errors import (
-    LLMBadRequestError as LLMBadRequestError,
-)
-from core.llm.errors import (
-    LLMConnectionError as LLMConnectionError,
-)
-from core.llm.errors import (
-    LLMInternalServerError as LLMInternalServerError,
-)
-from core.llm.errors import (
-    LLMRateLimitError as LLMRateLimitError,
-)
-from core.llm.errors import (
-    LLMTimeoutError as LLMTimeoutError,
-)
-from core.llm.fallback import MAX_RETRIES as MAX_RETRIES
-from core.llm.fallback import RETRY_BASE_DELAY as RETRY_BASE_DELAY
-from core.llm.fallback import RETRY_MAX_DELAY as RETRY_MAX_DELAY
-from core.llm.fallback import CircuitBreaker as CircuitBreaker
-from core.llm.fallback import retry_with_backoff_generic as retry_with_backoff_generic
-from core.llm.provider_dispatch import _cross_provider_dispatch as _cross_provider_dispatch
-from core.llm.provider_dispatch import _get_fallback_chain as _get_fallback_chain
-from core.llm.provider_dispatch import _get_provider_client as _get_provider_client
-from core.llm.provider_dispatch import _retry_provider_aware as _retry_provider_aware
-from core.llm.providers.anthropic import (
-    FALLBACK_MODELS as FALLBACK_MODELS,
-)
-from core.llm.providers.anthropic import (
-    NON_RETRYABLE_ERRORS as _NON_RETRYABLE_ERRORS,
-)
-from core.llm.providers.anthropic import (
-    RETRYABLE_ERRORS as _RETRYABLE_ERRORS,
-)
-from core.llm.providers.anthropic import (
-    _build_httpx_limits as _build_httpx_limits,
-)
-from core.llm.providers.anthropic import (
-    _build_httpx_timeout as _build_httpx_timeout,
-)
-from core.llm.providers.anthropic import (
-    get_anthropic_client as get_anthropic_client,
-)
-from core.llm.providers.anthropic import (
-    get_async_anthropic_client as get_async_anthropic_client,
-)
-from core.llm.providers.anthropic import (
-    reset_clients as reset_clients,
-)
-from core.llm.providers.anthropic import (
-    retry_with_backoff as _retry_with_backoff,
-)
-from core.llm.providers.anthropic import (
-    system_with_cache as _system_with_cache,
-)
-from core.llm.token_tracker import MODEL_PRICING as MODEL_PRICING
-from core.llm.token_tracker import LLMUsage as LLMUsage
-from core.llm.token_tracker import LLMUsageAccumulator as LLMUsageAccumulator
-from core.llm.token_tracker import calculate_cost as calculate_cost
-from core.llm.token_tracker import get_tracker
-from core.llm.token_tracker import get_usage_accumulator as get_usage_accumulator
-from core.llm.token_tracker import reset_usage_accumulator as reset_usage_accumulator
-from core.llm.token_tracker import track_token_usage as track_token_usage
-
-log = logging.getLogger(__name__)
-
-# ---------------------------------------------------------------------------
-# Hook system — optional lifecycle hooks for LLM call observability
-# ---------------------------------------------------------------------------
-
-_hooks_ctx: Any = None  # HookSystem | None — set via set_router_hooks()
-
-
-def set_router_hooks(hooks: Any) -> None:
-    """Wire HookSystem into the LLM router for LLM_CALL_START/END events.
-
-    Called from runtime wiring (bootstrap / pipeline_executor) after hooks are built.
-    """
-    global _hooks_ctx
-    _hooks_ctx = hooks
-
-
-def _fire_hook(event_name: str, data: dict[str, Any]) -> None:
-    """Fire a hook event if HookSystem is wired (or no-op)."""
-    fire_hook(_hooks_ctx, event_name, data)
-
-
-# Suppress langsmith/langchain rate-limit log spam (429 errors when quota exceeded)
-logging.getLogger("langsmith").setLevel(logging.ERROR)
-logging.getLogger("langchain").setLevel(logging.ERROR)
-
-
-# ---------------------------------------------------------------------------
-# LangSmith tracing (Phase 5-A): LangChain standard env vars
-# ---------------------------------------------------------------------------
-
-
-def is_langsmith_enabled() -> bool:
-    """Check if LangSmith tracing is active (both gate + key required).
-
-    Reads env vars at call time for testability.
-    """
-    tracing = _os.environ.get("LANGCHAIN_TRACING_V2", "").lower() == "true"
-    api_key = _os.environ.get("LANGCHAIN_API_KEY") or _os.environ.get("LANGSMITH_API_KEY")
-    return tracing and api_key is not None
-
-
-def maybe_traceable(
-    *,
-    run_type: str = "llm",
-    name: str | None = None,
-) -> Any:
-    """Return @traceable decorator if LangSmith is configured, else passthrough.
-
-    Public API — used by domain/verification layers for LangSmith integration.
-    """
-    if is_langsmith_enabled():
-        try:
-            from langsmith import traceable
-
-            return traceable(run_type=run_type, name=name)  # type: ignore[call-overload]
-        except ImportError:
-            pass
-
-    def _identity(fn: Any) -> Any:
-        return fn
-
-    return _identity
-
-
-# ---------------------------------------------------------------------------
-# Token usage recording helpers
-# ---------------------------------------------------------------------------
-
-
-def _record_response_usage(
-    response: Any,
-    model: str,
-    *,
-    label: str = "",
-) -> LLMUsage | None:
-    """Record token usage from an Anthropic response. Returns usage or None."""
-    if not (hasattr(response, "usage") and response.usage):
-        return None
-    in_tok = response.usage.input_tokens
-    out_tok = response.usage.output_tokens
-    cache_create = getattr(response.usage, "cache_creation_input_tokens", 0) or 0
-    cache_read = getattr(response.usage, "cache_read_input_tokens", 0) or 0
-    usage = get_tracker().record(
-        model,
-        in_tok,
-        out_tok,
-        cache_creation_tokens=cache_create,
-        cache_read_tokens=cache_read,
-    )
-    suffix = f" ({label})" if label else ""
-    log.info(
-        "LLM call%s: model=%s in=%d out=%d cost=$%.4f",
-        suffix,
-        model,
-        in_tok,
-        out_tok,
-        usage.cost_usd,
-    )
-    if cache_create or cache_read:
-        log.debug("Cache: create=%d read=%d", cache_create, cache_read)
-    return usage
-
-
-def _record_openai_usage(
-    response: Any,
-    model: str,
-    *,
-    label: str = "",
-) -> LLMUsage | None:
-    """Record token usage from an OpenAI-format response. Returns usage or None."""
-    if not (hasattr(response, "usage") and response.usage):
-        return None
-    in_tok = response.usage.prompt_tokens or 0
-    out_tok = response.usage.completion_tokens or 0
-    usage = get_tracker().record(model, in_tok, out_tok)
-    suffix = f" ({label})" if label else ""
-    log.info(
-        "LLM call%s: model=%s in=%d out=%d cost=$%.4f",
-        suffix,
-        model,
-        in_tok,
-        out_tok,
-        usage.cost_usd,
-    )
-    return usage
-
-
-# ---------------------------------------------------------------------------
-# call_with_failover — async model failover for AgenticLoop
-# ---------------------------------------------------------------------------
-
-
-async def call_with_failover(
-    models: list[str],
-    call_fn: Any,
-    *,
-    max_retries: int | None = None,
-    retry_base_delay: float | None = None,
-    retry_max_delay: float | None = None,
-) -> tuple[Any | None, str | None]:
-    """Execute an async LLM call with model failover chain.
-
-    Iterates through the ``models`` list. For each model, retries on
-    retryable errors (rate-limit, timeout, connection, server errors)
-    with exponential backoff. If all retries for a model are exhausted,
-    moves to the next model in the chain.
-
-    Non-retryable errors (e.g. AuthenticationError) cause immediate failure
-    without trying further models.
-
-    Args:
-        models: Ordered list of model names to try.
-        call_fn: Async callable ``(model: str) -> response``.
-        max_retries: Per-model retry count (default: settings.llm_max_retries).
-        retry_base_delay: Base delay in seconds (default: settings.llm_retry_base_delay).
-        retry_max_delay: Max delay cap in seconds (default: settings.llm_retry_max_delay).
-
-    Returns:
-        A tuple of ``(response, model_used)``. On complete failure,
-        returns ``(None, None)``.
-    """
-    import asyncio as _asyncio
-
-    _max_retries = max_retries if max_retries is not None else MAX_RETRIES
-    _base_delay = retry_base_delay if retry_base_delay is not None else RETRY_BASE_DELAY
-    _max_delay = retry_max_delay if retry_max_delay is not None else RETRY_MAX_DELAY
-
-    # Filter models by policy (GAP 2: model-policy.toml governance)
-    allowed_models = [m for m in models if is_model_allowed(m)]
-    if not allowed_models:
-        log.error("Failover: all models blocked by policy: %s", models)
-        return None, None
-
-    last_error: Exception | None = None
-    _t0_failover = time.monotonic()
-
-    for model_idx, current_model in enumerate(allowed_models):
-        for attempt in range(_max_retries):
-            try:
-                result = await call_fn(current_model)
-                return result, current_model
-            except _NON_RETRYABLE_ERRORS as exc:
-                # Context overflow → re-raise so adapter can set last_error
-                error_msg = str(exc).lower()
-                if "token" in error_msg or "context" in error_msg:
-                    log.warning(
-                        "Context overflow on model=%s: %s — propagating to adapter",
-                        current_model,
-                        type(exc).__name__,
-                    )
-                    raise
-                # Other non-retryable: re-raise so adapter can set last_error
-                log.warning(
-                    "Non-retryable error on model=%s: %s — propagating to adapter",
-                    current_model,
-                    type(exc).__name__,
-                )
-                raise
-            except _RETRYABLE_ERRORS as exc:
-                last_error = exc
-                wait = random.uniform(0, min(_base_delay * (2**attempt), _max_delay))
-                _elapsed = time.monotonic() - _t0_failover
-                log.debug(
-                    "Failover: model=%s attempt=%d/%d error=%s, retrying in %.1fs",
-                    current_model,
-                    attempt + 1,
-                    _max_retries,
-                    type(exc).__name__,
-                    wait,
-                )
-                # Emit retry_wait event for UX (elapsed timer + interrupt hint)
-                _fire_hook(
-                    "retry_wait",
-                    {
-                        "model": current_model,
-                        "attempt": attempt + 1,
-                        "max_retries": _max_retries,
-                        "delay_s": wait,
-                        "elapsed_s": _elapsed,
-                        "error_type": type(exc).__name__,
-                    },
-                )
-                if attempt < _max_retries - 1:
-                    await _asyncio.sleep(wait)
-            except Exception as exc:
-                # Unexpected error: log and move on to next model
-                last_error = exc
-                log.debug(
-                    "Failover: unexpected error on model=%s: %s",
-                    current_model,
-                    exc,
-                )
-                break  # break retry loop, try next model
-
-        # All retries exhausted for this model
-        if model_idx < len(allowed_models) - 1:
-            next_model = allowed_models[model_idx + 1]
-            log.info(
-                "Failover: model=%s exhausted, falling back to %s",
-                current_model,
-                next_model,
-            )
-
-    log.error(
-        "Failover: all models exhausted. Last error: %s",
-        last_error,
-    )
-    return None, None
-
-
-# ---------------------------------------------------------------------------
-# Data types
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class ToolCallRecord:
-    """Record of a single tool call within a tool-use loop."""
-
-    tool_name: str
-    tool_input: dict[str, Any]
-    tool_result: dict[str, Any]
-    duration_ms: float
-
-
-@dataclass
-class ToolUseResult:
-    """Result from a multi-turn tool-use LLM call."""
-
-    text: str
-    tool_calls: list[ToolCallRecord]
-    usage: list[LLMUsage]
-    rounds: int
-
-
-# ---------------------------------------------------------------------------
-# Main API functions — provider-aware routing
-# ---------------------------------------------------------------------------
-
-T = TypeVar("T", bound=BaseModel)
-
-
-def _route_provider(model: str) -> str:
-    """Resolve the provider for a model, honoring registered Plans.
-
-    v0.52.4 — wraps ``resolve_routing(model)`` so the LLM call dispatch
-    sees the actually-routed provider (e.g. ``openai-codex`` when a Plus
-    OAuth Plan is registered) instead of the static
-    ``_resolve_provider`` mapping (which always returned ``openai`` for
-    ``gpt-5.4`` regardless of OAuth state). Falls back to the static
-    resolver when no Plan is registered.
-
-    Caller flow unchanged: dispatch sees a provider string and looks up
-    the client via ``_get_provider_client(provider)``. The credential
-    lookup inside each provider client (``_get_codex_client`` etc.) now
-    consults the same ProfileStore, so the path is end-to-end coherent.
-    """
-    try:
-        from core.auth.plan_registry import resolve_routing
-
-        target = resolve_routing(model)
-        if target is not None:
-            return target.plan.provider
-    except Exception:
-        log.debug("resolve_routing failed for model=%s", model, exc_info=True)
-    return _resolve_provider(model)
-
-
-@maybe_traceable(run_type="llm", name="call_llm")  # type: ignore[untyped-decorator]
-def call_llm(
-    system: str,
-    user: str,
-    *,
-    model: str | None = None,
-    max_tokens: int = 4096,
-    temperature: float = 0.3,
-    seed: int | None = None,
-) -> str:
-    """Synchronous LLM call with provider-aware routing and failover.
-
-    Routes to Anthropic, OpenAI, or GLM SDK based on the model name.
-    Returns text content. Supports cross-provider fallback when enabled.
-    """
-    target_model = model or settings.model
-    provider = _route_provider(target_model)
-
-    if seed is not None:
-        log.info("Reproducibility seed=%d requested (logged for auditing)", seed)
-
-    def _dispatch(p: str, m: str) -> str:
-        _fire_hook(
-            "llm_call_start",
-            {"model": m, "provider": p, "function": "call_llm"},
-        )
-        t0 = time.monotonic()
-        try:
-            if p != "anthropic":
-                oa_client = _get_provider_client(p)
-
-                def _do_call_openai(*, model: str) -> str:
-                    response = oa_client.chat.completions.create(
-                        model=model,
-                        max_completion_tokens=max_tokens,
-                        temperature=temperature,
-                        messages=[
-                            {"role": "system", "content": system},
-                            {"role": "user", "content": user},
-                        ],
-                        timeout=120.0,
-                    )
-                    _record_openai_usage(response, model)
-                    choice = response.choices[0]
-                    return choice.message.content or ""
-
-                result: str = _retry_provider_aware(_do_call_openai, model=m, provider=p)
-            else:
-                client = get_anthropic_client()
-                system_cached = _system_with_cache(system)
-
-                def _do_call(*, model: str) -> str:
-                    response = client.messages.create(
-                        model=model,
-                        max_tokens=max_tokens,
-                        temperature=temperature,
-                        system=system_cached,
-                        messages=[{"role": "user", "content": user}],
-                    )
-                    _record_response_usage(response, model)
-
-                    block = response.content[0]
-                    if not hasattr(block, "text"):
-                        raise TypeError(f"Expected TextBlock, got {type(block)}")
-                    return block.text
-
-                result = _retry_with_backoff(_do_call, model=m)
-        except Exception as exc:
-            elapsed_ms = (time.monotonic() - t0) * 1000
-            _fire_hook(
-                "llm_call_end",
-                {
-                    "model": m,
-                    "provider": p,
-                    "function": "call_llm",
-                    "latency_ms": elapsed_ms,
-                    "error": str(exc),
-                },
-            )
-            raise
-
-        elapsed_ms = (time.monotonic() - t0) * 1000
-        _fire_hook(
-            "llm_call_end",
-            {
-                "model": m,
-                "provider": p,
-                "function": "call_llm",
-                "latency_ms": elapsed_ms,
-                "error": None,
-            },
-        )
-        return result
-
-    return _cross_provider_dispatch(provider, target_model, _dispatch, "call_llm")
-
-
-@maybe_traceable(run_type="llm", name="call_llm_parsed")  # type: ignore[untyped-decorator]
-def call_llm_parsed(  # noqa: UP047 — PEP695 syntax requires Python 3.12+
-    system: str,
-    user: str,
-    *,
-    output_model: type[T],
-    model: str | None = None,
-    max_tokens: int = 4096,
-    temperature: float = 0.3,
-) -> T:
-    """LLM call with provider-aware structured output routing.
-
-    Supports cross-provider fallback when enabled.
-    """
-    target_model = model or settings.model
-    provider = _route_provider(target_model)
-
-    def _dispatch(p: str, m: str) -> T:
-        _fire_hook(
-            "llm_call_start",
-            {"model": m, "provider": p, "function": "call_llm_parsed"},
-        )
-        t0 = time.monotonic()
-        try:
-            if p != "anthropic":
-                oa_client = _get_provider_client(p)
-
-                def _do_call_openai(*, model: str) -> T:
-                    response = oa_client.beta.chat.completions.parse(
-                        model=model,
-                        max_completion_tokens=max_tokens,
-                        temperature=temperature,
-                        messages=[
-                            {"role": "system", "content": system},
-                            {"role": "user", "content": user},
-                        ],
-                        response_format=output_model,
-                        timeout=120.0,
-                    )
-                    _record_openai_usage(response, model, label="parsed")
-
-                    choice = response.choices[0]
-                    if choice.message.parsed is None:
-                        raise ValueError(
-                            "LLM returned no structured output. "
-                            "Verify the prompt constrains the response format "
-                            "and the Pydantic model matches the schema."
-                        )
-                    return choice.message.parsed  # type: ignore[no-any-return]
-
-                result: T = _retry_provider_aware(_do_call_openai, model=m, provider=p)
-            else:
-                client = get_anthropic_client()
-                system_cached = _system_with_cache(system)
-
-                def _do_call(*, model: str) -> T:
-                    response = client.messages.parse(
-                        model=model,
-                        max_tokens=max_tokens,
-                        temperature=temperature,
-                        system=system_cached,
-                        messages=[{"role": "user", "content": user}],
-                        output_format=output_model,
-                    )
-                    _record_response_usage(response, model, label="parsed")
-
-                    if response.parsed_output is None:
-                        raise ValueError(
-                            "LLM returned no structured output. "
-                            "Verify the prompt constrains the response format "
-                            "and the Pydantic model matches the schema."
-                        )
-                    return response.parsed_output
-
-                result = _retry_with_backoff(_do_call, model=m)
-        except Exception as exc:
-            elapsed_ms = (time.monotonic() - t0) * 1000
-            _fire_hook(
-                "llm_call_end",
-                {
-                    "model": m,
-                    "provider": p,
-                    "function": "call_llm_parsed",
-                    "latency_ms": elapsed_ms,
-                    "error": str(exc),
-                },
-            )
-            raise
-
-        elapsed_ms = (time.monotonic() - t0) * 1000
-        _fire_hook(
-            "llm_call_end",
-            {
-                "model": m,
-                "provider": p,
-                "function": "call_llm_parsed",
-                "latency_ms": elapsed_ms,
-                "error": None,
-            },
-        )
-        return result
-
-    return _cross_provider_dispatch(provider, target_model, _dispatch, "call_llm_parsed")
-
-
-@maybe_traceable(run_type="llm", name="call_llm_json")  # type: ignore[untyped-decorator]
-def call_llm_json(
-    system: str,
-    user: str,
-    *,
-    model: str | None = None,
-    max_tokens: int = 4096,
-    temperature: float = 0.3,
-    seed: int | None = None,
-) -> dict[str, Any]:
-    """Claude call that parses JSON from the response. Includes failover.
-
-    Legacy function — prefer call_llm_parsed() for guaranteed structured output.
-    Kept for backward compatibility with code that expects dict[str, Any].
-    """
-    raw = call_llm(
-        system, user, model=model, max_tokens=max_tokens, temperature=temperature, seed=seed
-    )
-    # Strip markdown code fences if present (handles ```json, ``` with trailing spaces, etc.)
-    text = raw.strip()
-    text = re.sub(r"^```\w*\s*\n", "", text)
-    text = re.sub(r"\n```\s*$", "", text)
-    text = text.strip()
-
-    # Try direct JSON parse
-    try:
-        result: dict[str, Any] = json.loads(text)
-        return result
-    except json.JSONDecodeError:
-        pass
-
-    # Fallback: extract JSON object from text (handles markdown-wrapped responses)
-    first_brace = text.find("{")
-    last_brace = text.rfind("}")
-    if first_brace != -1 and last_brace > first_brace:
-        candidate = text[first_brace : last_brace + 1]
-        try:
-            result = json.loads(candidate)
-            log.info("Extracted JSON from position %d-%d in LLM response", first_brace, last_brace)
-            return result
-        except json.JSONDecodeError:
-            pass
-
-    log.error("Failed to parse LLM JSON response. Raw text: %s", text[:500])
-    raise ValueError("LLM returned invalid JSON: could not extract JSON object from response")
-
-
-@maybe_traceable(run_type="chain", name="call_llm_with_tools")  # type: ignore[untyped-decorator]
-def call_llm_with_tools(
-    system: str,
-    user: str,
-    *,
-    tools: list[dict[str, Any]],
-    tool_executor: Any,
-    model: str | None = None,
-    max_tokens: int = 4096,
-    temperature: float = 0.3,
-    max_tool_rounds: int = 5,
-) -> ToolUseResult:
-    """LLM call with tool-use loop and provider-aware routing.
-
-    Supports cross-provider fallback when enabled.
-    """
-    target_model = model or settings.model
-    provider = _route_provider(target_model)
-
-    def _dispatch(p: str, m: str) -> ToolUseResult:
-        _fire_hook(
-            "llm_call_start",
-            {"model": m, "provider": p, "function": "call_llm_with_tools"},
-        )
-        t0_tools = time.monotonic()
-
-        try:
-            if p != "anthropic":
-                from core.llm.providers.openai import OpenAIAdapter
-
-                adapter = OpenAIAdapter(default_model=m)
-                if p == "glm":
-                    from core.llm.providers import openai as _openai_provider
-
-                    orig_get = _openai_provider._get_openai_client
-
-                    def _glm_client_override() -> Any:
-                        return _get_provider_client("glm")
-
-                    _openai_provider._get_openai_client = _glm_client_override
-                    try:
-                        oai_result: ToolUseResult = adapter.generate_with_tools(
-                            system,
-                            user,
-                            tools=tools,
-                            tool_executor=tool_executor,
-                            model=m,
-                            max_tokens=max_tokens,
-                            temperature=temperature,
-                            max_tool_rounds=max_tool_rounds,
-                        )
-                    finally:
-                        _openai_provider._get_openai_client = orig_get
-                    tools_result = oai_result
-                else:
-                    oai_result2: ToolUseResult = adapter.generate_with_tools(
-                        system,
-                        user,
-                        tools=tools,
-                        tool_executor=tool_executor,
-                        model=m,
-                        max_tokens=max_tokens,
-                        temperature=temperature,
-                        max_tool_rounds=max_tool_rounds,
-                    )
-                    tools_result = oai_result2
-            else:
-                client = get_anthropic_client()
-                system_cached = _system_with_cache(system)
-
-                all_tool_calls: list[ToolCallRecord] = []
-                all_usage: list[LLMUsage] = []
-                messages: list[dict[str, Any]] = [{"role": "user", "content": user}]
-
-                cached_tools: list[dict[str, Any]] = []
-                for i, tool in enumerate(tools):
-                    tool_copy = dict(tool)
-                    if i == len(tools) - 1:
-                        tool_copy["cache_control"] = {"type": "ephemeral"}
-                    cached_tools.append(tool_copy)
-                tools_for_api = cached_tools
-
-                for round_idx in range(max_tool_rounds):
-                    is_last_round = round_idx == max_tool_rounds - 1
-                    tool_choice: dict[str, str] | None = (
-                        {"type": "none"} if is_last_round else {"type": "auto"}
-                    )
-
-                    def _do_call(
-                        *,
-                        model: str,
-                        _tc: dict[str, str] | None = tool_choice,
-                    ) -> Any:
-                        return client.messages.create(  # type: ignore[call-overload]
-                            model=model,
-                            max_tokens=max_tokens,
-                            temperature=temperature,
-                            system=system_cached,
-                            messages=messages,
-                            tools=tools_for_api,
-                            tool_choice=_tc,
-                        )
-
-                    response = _retry_with_backoff(_do_call, model=m)
-
-                    usage = _record_response_usage(response, m, label="tools")
-                    if usage is not None:
-                        all_usage.append(usage)
-
-                    if response.stop_reason != "tool_use":
-                        text = ""
-                        for block in response.content:
-                            if hasattr(block, "text"):
-                                text += block.text
-                        tools_result = ToolUseResult(
-                            text=text,
-                            tool_calls=all_tool_calls,
-                            usage=all_usage,
-                            rounds=round_idx + 1,
-                        )
-                        break
-
-                    assistant_content = response.content
-                    tool_result_blocks: list[dict[str, Any]] = []
-
-                    for block in assistant_content:
-                        if block.type != "tool_use":
-                            continue
-                        tool_name = block.name
-                        tool_input = block.input
-                        t0_tool = time.time()
-                        try:
-                            tool_output: dict[str, Any] = tool_executor(tool_name, **tool_input)
-                        except Exception as exc:
-                            log.warning("Tool '%s' execution failed: %s", tool_name, exc)
-                            from core.auth.scrub import scrub_credentials
-
-                            tool_output = {"error": scrub_credentials(str(exc))}
-                        elapsed_ms_tool = (time.time() - t0_tool) * 1000
-
-                        all_tool_calls.append(
-                            ToolCallRecord(
-                                tool_name=tool_name,
-                                tool_input=tool_input,
-                                tool_result=tool_output,
-                                duration_ms=elapsed_ms_tool,
-                            )
-                        )
-                        tool_result_blocks.append(
-                            {
-                                "type": "tool_result",
-                                "tool_use_id": block.id,
-                                "content": json.dumps(tool_output),
-                            }
-                        )
-
-                    messages.append({"role": "assistant", "content": assistant_content})
-                    messages.append({"role": "user", "content": tool_result_blocks})
-                else:
-                    tools_result = ToolUseResult(
-                        text="",
-                        tool_calls=all_tool_calls,
-                        usage=all_usage,
-                        rounds=max_tool_rounds,
-                    )
-        except Exception as exc:
-            elapsed_ms_total = (time.monotonic() - t0_tools) * 1000
-            _fire_hook(
-                "llm_call_end",
-                {
-                    "model": m,
-                    "provider": p,
-                    "function": "call_llm_with_tools",
-                    "latency_ms": elapsed_ms_total,
-                    "error": str(exc),
-                },
-            )
-            raise
-
-        elapsed_ms_total = (time.monotonic() - t0_tools) * 1000
-        _fire_hook(
-            "llm_call_end",
-            {
-                "model": m,
-                "provider": p,
-                "function": "call_llm_with_tools",
-                "latency_ms": elapsed_ms_total,
-                "error": None,
-            },
-        )
-        return tools_result
-
-    return _cross_provider_dispatch(provider, target_model, _dispatch, "call_llm_with_tools")
-
-
-@maybe_traceable(run_type="llm", name="call_llm_streaming")  # type: ignore[untyped-decorator]
-def call_llm_streaming(
-    system: str,
-    user: str,
-    *,
-    model: str | None = None,
-    max_tokens: int = 4096,
-    temperature: float = 0.3,
-) -> Iterator[str]:
-    """Streaming Claude call with failover. Yields text deltas."""
-    from core.llm.providers.anthropic import get_circuit_breaker
-
-    client = get_anthropic_client()
-    target_model = model or settings.model
-    provider = _route_provider(target_model)
-    circuit_breaker = get_circuit_breaker()
-
-    # Hook: LLM_CALL_START
-    _fire_hook(
-        "llm_call_start",
-        {"model": target_model, "provider": provider, "function": "call_llm_streaming"},
-    )
-    t0 = time.monotonic()
-
-    def _do_stream(*, model: str) -> Iterator[str]:
-        system_cached = _system_with_cache(system)
-        with client.messages.stream(
-            model=model,
-            max_tokens=max_tokens,
-            temperature=temperature,
-            system=system_cached,
-            messages=[{"role": "user", "content": user}],
-        ) as stream:
-            yield from stream.text_stream
-
-            # Record success AFTER stream completes
-            circuit_breaker.record_success()
-
-            # Capture token usage from the stream's final response
-            final = stream.get_final_message()
-            if final:
-                _record_response_usage(final, model, label="stream")
-
-    def _hooked_stream(inner: Iterator[str]) -> Iterator[str]:
-        """Wrapper that fires LLM_CALL_END after stream exhaustion or error."""
-        try:
-            yield from inner
-        except Exception as exc:
-            elapsed_ms = (time.monotonic() - t0) * 1000
-            _fire_hook(
-                "llm_call_end",
-                {
-                    "model": target_model,
-                    "provider": provider,
-                    "function": "call_llm_streaming",
-                    "latency_ms": elapsed_ms,
-                    "error": str(exc),
-                },
-            )
-            raise
-        else:
-            elapsed_ms = (time.monotonic() - t0) * 1000
-            _fire_hook(
-                "llm_call_end",
-                {
-                    "model": target_model,
-                    "provider": provider,
-                    "function": "call_llm_streaming",
-                    "latency_ms": elapsed_ms,
-                    "error": None,
-                },
-            )
-
-    # Circuit breaker check
-    if not circuit_breaker.can_execute():
-        raise RuntimeError(
-            "Circuit breaker is open — LLM API calls are temporarily blocked. "
-            "Too many consecutive failures detected."
-        )
-
-    # For streaming, retry at connection level only
-    models_to_try = [target_model] + [m for m in FALLBACK_MODELS if m != target_model]
-    last_error: Exception | None = None
-
-    for current_model in models_to_try:
-        for attempt in range(MAX_RETRIES):
-            try:
-                result = _do_stream(model=current_model)
-                return _hooked_stream(result)
-            except _RETRYABLE_ERRORS as exc:
-                last_error = exc
-                delay = min(RETRY_BASE_DELAY * (2**attempt), RETRY_MAX_DELAY)
-                log.warning(
-                    "Streaming call failed (model=%s, attempt=%d/%d): %s",
-                    current_model,
-                    attempt + 1,
-                    MAX_RETRIES,
-                    type(exc).__name__,
-                )
-                time.sleep(delay)
-
-    if last_error is None:
-        raise RuntimeError("All retries exhausted with no error recorded")
-    circuit_breaker.record_failure()
-    raise last_error
-
-
-# ---------------------------------------------------------------------------
-# Thread-safe injection via contextvars
-# ---------------------------------------------------------------------------
-
-_llm_json_ctx: ContextVar[LLMJsonCallable | None] = ContextVar("llm_json", default=None)
-_llm_parsed_ctx: ContextVar[LLMParsedCallable | None] = ContextVar("llm_parsed", default=None)
-_llm_tool_ctx: ContextVar[LLMToolCallable | None] = ContextVar("llm_tool", default=None)
-
-# Secondary LLM contextvars for ensemble/cross-LLM mode
-_secondary_llm_json_ctx: ContextVar[LLMJsonCallable | None] = ContextVar(
-    "secondary_llm_json", default=None
-)
-_secondary_llm_parsed_ctx: ContextVar[LLMParsedCallable | None] = ContextVar(
-    "secondary_llm_parsed", default=None
-)
-
-
-def set_llm_callable(
-    json_fn: LLMJsonCallable,
-    text_fn: LLMTextCallable,
-    parsed_fn: LLMParsedCallable | None = None,
-    tool_fn: LLMToolCallable | None = None,
-    secondary_json_fn: LLMJsonCallable | None = None,
-    secondary_parsed_fn: LLMParsedCallable | None = None,
-) -> None:
-    """Inject LLM callables (typically called by GeodeRuntime.create())."""
-    _llm_json_ctx.set(json_fn)
-    if parsed_fn is not None:
-        _llm_parsed_ctx.set(parsed_fn)
-    if tool_fn is not None:
-        _llm_tool_ctx.set(tool_fn)
-    # Always update secondary contextvars (set to None to clear if not provided)
-    _secondary_llm_json_ctx.set(secondary_json_fn)
-    _secondary_llm_parsed_ctx.set(secondary_parsed_fn)
-
-
-def get_llm_json() -> LLMJsonCallable:
-    """Return the injected JSON callable. Raises if not injected."""
-    fn = _llm_json_ctx.get()
-    if fn is None:
-        raise RuntimeError(
-            "LLM JSON callable not injected. "
-            "Call set_llm_callable() first (done by GeodeRuntime.create())."
-        )
-    return fn
-
-
-def get_llm_parsed() -> LLMParsedCallable:
-    """Return the injected parsed callable. Raises if not injected."""
-    fn = _llm_parsed_ctx.get()
-    if fn is None:
-        raise RuntimeError(
-            "LLM parsed callable not injected. "
-            "Call set_llm_callable(parsed_fn=...) first (done by GeodeRuntime.create())."
-        )
-    return fn
-
-
-def get_llm_tool() -> LLMToolCallable:
-    """Return the injected tool-use callable. Raises if not injected."""
-    fn = _llm_tool_ctx.get()
-    if fn is None:
-        raise RuntimeError(
-            "LLM tool callable not injected. "
-            "Call set_llm_callable(tool_fn=...) first (done by GeodeRuntime.create())."
-        )
-    return fn
-
-
-def get_secondary_llm_json() -> LLMJsonCallable | None:
-    """Return the secondary JSON callable, or None if not configured."""
-    return _secondary_llm_json_ctx.get()
-
-
-def get_secondary_llm_parsed() -> LLMParsedCallable | None:
-    """Return the secondary parsed callable, or None if not configured."""
-    return _secondary_llm_parsed_ctx.get()
