+++ tests/test_failover.py
@@ -77,7 +77,7 @@ class TestRetryWithBackoff:
 
 
 class TestCallLlmFailover:
-    @patch("core.llm.router.get_anthropic_client")
+    @patch("core.llm.router.calls.text.get_anthropic_client")
     def test_call_llm_success(self, mock_client_fn):
         mock_response = MagicMock()
         mock_block = MagicMock()
@@ -89,7 +89,7 @@ class TestCallLlmFailover:
         assert result == "response text"
 
     @patch("core.llm.fallback.time.sleep")
-    @patch("core.llm.router.get_anthropic_client")
+    @patch("core.llm.router.calls.text.get_anthropic_client")
     def test_call_llm_retries_on_failure(self, mock_client_fn, mock_sleep):
         rate_err = anthropic.RateLimitError.__new__(anthropic.RateLimitError)
         rate_err.status_code = 429
@@ -117,19 +117,19 @@ class TestFallbackModels:
 class TestCallLlmJsonExtraction:
     """Tests for robust JSON extraction from LLM responses."""
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_direct_json(self, mock_call):
         mock_call.return_value = '{"key": "value"}'
         result = call_llm_json("sys", "usr")
         assert result == {"key": "value"}
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_json_with_markdown_fences(self, mock_call):
         mock_call.return_value = '```json\n{"key": "value"}\n```'
         result = call_llm_json("sys", "usr")
         assert result == {"key": "value"}
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_json_embedded_in_text(self, mock_call):
         """LLM returns prose before/after JSON — extraction finds the object."""
         mock_call.return_value = (
@@ -142,7 +142,7 @@ class TestCallLlmJsonExtraction:
         assert result["evaluator_type"] == "quality_judge"
         assert result["composite_score"] == 72.0
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_json_with_nested_braces(self, mock_call):
         mock_call.return_value = (
             'Some text\n{"axes": {"d_score": 4.0, "e_score": 3.0}, "rationale": "test"}\nMore text'
@@ -150,13 +150,13 @@ class TestCallLlmJsonExtraction:
         result = call_llm_json("sys", "usr")
         assert result["axes"]["d_score"] == 4.0
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_no_json_raises_error(self, mock_call):
         mock_call.return_value = "This is a plain text response with no JSON."
         with pytest.raises(ValueError, match="invalid JSON"):
             call_llm_json("sys", "usr")
 
-    @patch("core.llm.router.call_llm")
+    @patch("core.llm.router.calls.json.call_llm")
     def test_whitespace_around_json(self, mock_call):
         mock_call.return_value = '  \n  {"key": 42}  \n  '
         result = call_llm_json("sys", "usr")
